Series Palette Generator.

This function builds palettes based on popular series and anime from Netflix and others visualmedia. The selection has been subjectively made by the authors.

Usage

serie_palette(
  name,
  palette_family = palette_netflyx,
  n,
  type = c("discrete", "continuous")
)

Arguments

name

Name of the color palette. Current options are Netflyx and Anime.

palette_family

list of palettes to use.

n

Number of colors to be used. Each palette includes up to 9 colors for discrete palettes. Any n can be chosen for continuous palettes.

type

Usage of palette as "continuous" or "discrete". Continuous palettes interpolate the three first colors of the palette to create a gradient. If not specified, function assumes continuous if n>9 and discrete if n<9.

Value

A vector of n colors